Modifie types_contrat en contrat_de_travail_type et supprime TypesContrat(Enum)
